backend/ragai/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.1
python-multipart==0.0.9
PyPDF2==3.0.1
numpy==1.26.4
sentence-transformers==2.5.1
faiss-cpu==1.7.4
tiktoken==0.6.0 